PointArrays

PointArrays

RealNeuralNetworks.NodeNets.PointArrays.add_offset! — Method.

add offset to points

source

RealNeuralNetworks.NodeNets.PointArrays.from_binary_image — Method.

parameter: bin_im: binary array. object voxels are false, non-object voxels are true!

source

RealNeuralNetworks.NodeNets.PointArrays.from_seg — Method.

find points inside an object from a segmentation array.

source

RealNeuralNetworks.NodeNets.PointArrays.get_boundary_point_indexes — Method.

find out the boundary voxels and represent them as indexes in the point array

source

RealNeuralNetworks.NodeNets.PointArrays.merge — Method.

merge(self::Array{T,3}, other::Array{T,3})

source